+++ src/Analysis/Source/Visitors/SymbolCollector.php
@@ -7,6 +7,7 @@ namespace Cluion\Moduark\Analysis\Source\Visitors;
 use Cluion\Moduark\Analysis\Source\SourceSymbol;
 use Cluion\Moduark\Module;
 use PhpParser\Node;
+use PhpParser\Node\Stmt\Class_;
 use PhpParser\Node\Stmt\ClassLike;
 use PhpParser\NodeVisitorAbstract;
 
@@ -35,6 +36,9 @@ final class SymbolCollector extends NodeVisitorAbstract
             $this->owner,
             $this->file,
             $node->getStartLine(),
+            $node instanceof Class_ && $node->extends !== null
+                ? $node->extends->toString()
+                : null,
         );
 
         return null;