+++ tests/Fixtures/LevelThree/Modules/User/Models/BaseModel.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\LevelThree\Modules\User\Models;
+
+use Illuminate\Database\Eloquent\Model as EloquentModel;
+
+abstract class BaseModel extends EloquentModel
+{
+}