+++ tests/Fixtures/LevelThree/Modules/User/UserModule.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Fixtures\LevelThree\Modules\User;
+
+use Cluion\Moduark\Module;
+
+final class UserModule extends Module
+{
+}